+++ core/src/test/java/nl/nn/adapterframework/senders/XsltSenderTest.java
@@ -19,7 +19,23 @@ public class XsltSenderTest extends SenderTestBase<XsltSender> {
 	}
 
 	@Test
-	public void basic() throws SenderException, TimeOutException, ConfigurationException, IOException {
+	public void basicXslt1() throws SenderException, TimeOutException, ConfigurationException, IOException {
+		sender.setStyleSheetName("/Xslt/duplicateImport/root.xsl");
+		sender.setXsltVersion(1);
+		sender.configure();
+		sender.open();
+		String input=getFile("/Xslt/duplicateImport/in.xml");
+		log.debug("inputfile ["+input+"]");
+		String expected=getFile("/Xslt/duplicateImport/out.xml");
+
+		ParameterResolutionContext prc = new ParameterResolutionContext(input, session);
+		String result = sender.sendMessage(null, input, prc);
+
+		assertEquals(expected.replaceAll("\\s",""), result.replaceAll("\\s",""));
+	}
+
+	@Test
+	public void basicXslt2() throws SenderException, TimeOutException, ConfigurationException, IOException {
 		sender.setStyleSheetName("/Xslt/duplicateImport/root.xsl");
 		sender.configure();
 		sender.open();
@@ -31,5 +47,48 @@ public class XsltSenderTest extends SenderTestBase<XsltSender> {
 		String result = sender.sendMessage(null, input, prc);
 
 		assertEquals(expected.replaceAll("\\s",""), result.replaceAll("\\s",""));
+	}
+
+	@Test
+	public void testSimpleXslt1Xpath() throws SenderException, TimeOutException, ConfigurationException, IOException {
+		sender.setXpathExpression("result");
+		sender.setXsltVersion(1);
+		sender.configure();
+		sender.open();
+
+		String input = "<result>dummy</result>";
+		ParameterResolutionContext prc = new ParameterResolutionContext(input, session);
+		String result = sender.sendMessage(null, input, prc);
+
+		assertEquals("dummy", result);
+	}
+
+	@Test
+	public void testSimpleXslt2Xpath() throws SenderException, TimeOutException, ConfigurationException, IOException {
+		sender.setXpathExpression("result");
+		sender.configure();
+		sender.open();
+
+		String input = "<result>dummy</result>";
+		ParameterResolutionContext prc = new ParameterResolutionContext(input, session);
+		String result = sender.sendMessage(null, input, prc);
+
+		assertEquals("dummy", result);
+	}
+
+	//This xPath only runs on xslt version 1
+	@Test
+	public void testComplexXslt1Xpath() throws SenderException, TimeOutException, ConfigurationException, IOException {
+		sender.setXpathExpression("number(count(/results/result[contains(@name , 'test')]))");
+		sender.setOutputType("txt");
+		sender.setXsltVersion(1);
+		sender.configure();
+		sender.open();
+
+		String input = "<results><result name='test1'>dummy1</result><result name='test2'>dummy2</result></results>";
+		ParameterResolutionContext prc = new ParameterResolutionContext(input, session);
+		String result = sender.sendMessage(null, input, prc);
+
+		assertEquals("2", result);
 	}
 }