+++ sap/src/main/java/nl/nn/adapterframework/extensions/sap/jco2/SapSenderBase.java
@@ -20,6 +20,7 @@ import nl.nn.adapterframework.core.IPipeLineSession;
 import nl.nn.adapterframework.core.ISenderWithParameters;
 import nl.nn.adapterframework.core.SenderException;
 import nl.nn.adapterframework.core.TimeOutException;
+import nl.nn.adapterframework.extensions.sap.SapException;
 import nl.nn.adapterframework.extensions.sap.jco2.tx.ClientFactoryUtils;
 import nl.nn.adapterframework.parameters.Parameter;
 import nl.nn.adapterframework.parameters.ParameterList;
@@ -61,6 +62,7 @@ public abstract class SapSenderBase extends SapFunctionFacade implements ISender
 	protected ParameterList paramList = null;
 	
 	
+	@Override
 	public void configure() throws ConfigurationException {
 		super.configure();
 		if (paramList!=null) {
@@ -79,6 +81,7 @@ public abstract class SapSenderBase extends SapFunctionFacade implements ISender
 		}
 	}
 
+	@Override
 	public void open() throws SenderException {
 		try {
 			openFacade();
@@ -88,10 +91,12 @@ public abstract class SapSenderBase extends SapFunctionFacade implements ISender
 		}
 	}
 	
+	@Override
 	public void close() {
 		closeFacade();
 	}
 
+	@Override
 	public String sendMessage(String correlationID, String message) throws SenderException, TimeOutException {
 		return sendMessage(correlationID,message,null);
 	}
@@ -152,6 +157,7 @@ public abstract class SapSenderBase extends SapFunctionFacade implements ISender
 		return ClientFactoryUtils.getTransactionalTid(sapSystem,client,true);
 	}
 
+	@Override
 	public void addParameter(Parameter p) {
 		if (paramList==null) {
 			paramList=new ParameterList();
@@ -178,6 +184,7 @@ public abstract class SapSenderBase extends SapFunctionFacade implements ISender
 	protected void setSynchronous(boolean b) {
 		synchronous = b;
 	}
+	@Override
 	public boolean isSynchronous() {
 		return synchronous;
 	}