+++ sap/src/main/java/nl/nn/adapterframework/extensions/sap/jco3/handlers/TableHandler.java
@@ -42,10 +42,11 @@ public class TableHandler extends Handler {
 		this.table=table;
 	}
 
+	@Override
 	protected void startElement(String localName) {
 		if (!parsedItem) {
 			if (localName.equals("item")) {
-				log.debug("appending row");
+				if(log.isTraceEnabled()) log.trace("appending row");
 				table.appendRow();
 				parsedItem=true;
 			} else {
@@ -65,6 +66,7 @@ public class TableHandler extends Handler {
 		}
 	}
 
+	@Override
 	protected void endElement(String localName) {
 		if (parsedStringField) {
 			endStringField(localName, table);